re-enabled and simplified SampleCheckpoint, tweaked subdir and moved checkpoints back to subdir
